Only image MS files

+++ katsdpcal/scripts/sim_l1_receive.py
@@ -72,7 +72,7 @@ def accumulate_l1(rx, return_data=False):
 if __name__ == '__main__':
     """
     Recieve an L1 output stream and print some details to confirm all is going well
-    Optionally write the L1 data back to the h5 file
+    Optionally write the L1 data back to the MS or H5 file
     """
     opts = parse_opts() 
     # Initialise spead receiver
@@ -94,18 +94,16 @@ if __name__ == '__main__':
         ts = opts.telstate
 
         # was our simulator using an H5 or MS file?
-        file_class = get_file_format(opts.file)
+        #file_class = get_file_format(opts.file)
 
         if not ts.cal_full_l1:
             print 'Only target L1 stream transmitted. Not saving L1 data to file.'
-        elif file_class != SimDataMS:
-            print 'Simulator didnt use MS file. Can only save L1 data to MS, so not saving L1 data to file.'
         else:
             if os.path.isfile(new_file) or os.path.isdir(new_file):
                 print 'WARNING: L1 data file {0} already exists. Over writing it.'.format(new_file,)
-                shutil.rmtree(new_file)
+                os.system('rm -rf {0}'.format(new_file,))
 
-            shutil.copytree(opts.file,new_file)
+            os.system('cp -r {0} {1}'.format(opts.file,new_file))
             # set up file to write the data into
             datafile = init_simdata(new_file,mode='r+')
 
@@ -114,23 +112,28 @@ if __name__ == '__main__':
             datafile.close()
 
             if opts.image:
-                bchan = ts.cal_bchan
-                echan = ts.cal_echan
+                # only image from MS simulator
+                if file_class == SimDataMS:
+                    bchan = ts.cal_bchan
+                    echan = ts.cal_echan-ts.cal_bchan
 
-                # image L0 data
-                if os.path.isfile(new_file) or os.path.isdir(new_file):
-                    print 'WARNING: L0 image L0_I* already exists. Over writing it.'
-                    os.system('rm -rf L0_I*')
+                    # image L0 data
+                    if os.path.isfile(new_file) or os.path.isdir(new_file):
+                        print 'WARNING: L0 image L0_I* already exists. Over writing it.'
+                        os.system('rm -rf L0_I*')
 
-                # image using casapy
-                clean_params = 'vis="{0}",imagename="L0_I",niter=0,stokes="I",spw="0:{1}~{2}",field="0",cell="30arcsec"'.format(opts.file,bchan,echan)
-                os.system("casapy -c 'clean({0})' ".format(clean_params))
+                    # image using casapy
+                    clean_params = 'vis="{0}",imagename="L0_I",niter=0,stokes="I",spw="0:{1}~{2}",field="0",cell="30arcsec"'.format(opts.file,bchan,echan)
+                    os.system("casapy -c 'clean({0})' ".format(clean_params))
 
-                # image L1 data
-                if os.path.isfile(new_file) or os.path.isdir(new_file):
-                    print 'WARNING: L1 image L1_I* already exists. Over writing it.'
-                    os.system('rm -rf L1_I*')
+                    # image L1 data
+                    if os.path.isfile(new_file) or os.path.isdir(new_file):
+                        print 'WARNING: L1 image L1_I* already exists. Over writing it.'
+                        os.system('rm -rf L1_I*')
 
-                # image using casapy
-                clean_params = 'vis="{0}",imagename="L1_I",niter=0,stokes="I",spw="0:{1}~{2}",field="0",cell="30arcsec"'.format(new_file,bchan,echan)
-                os.system("casapy -c 'clean({0})' ".format(clean_params))
+                    # image using casapy
+                    clean_params = 'vis="{0}",imagename="L1_I",niter=0,stokes="I",spw="0:{1}~{2}",field="0",cell="30arcsec"'.format(new_file,bchan,echan)
+                    os.system("casapy -c 'clean({0})' ".format(clean_params))
+
+                else:
+                    print 'Simulator didnt use MS file. Can only currently image from MS.'
